requirements.txt
Django~=5.2.3
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
gunicorn>=21.2,<22.0
pillow>=10.0,<11.0
python-decouple>=3.8,<4.0
python-dotenv>=1.2.1,<2.0
dj-database-url>=1.0,<2.0
cloudinary>=1.32,<2.0
django-cloudinary-storage>=0.3,<0.4
whitenoise>=6.5,<7.0
psycopg2-binary>=2.9,<3.0
